+++ src/main/java/sistemaDeVendas/services/ClienteService.java
@@ -1,11 +1,17 @@
 package sistemaDeVendas.services;
 
 import java.util.ArrayList;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import sistemaDeVendas.domains.Cliente;
+import sistemaDeVendas.exceptions.CriacaoInvalidaException;
+import sistemaDeVendas.exceptions.IdInvalidoException;
+import sistemaDeVendas.exceptions.NotFoundException;
+import sistemaDeVendas.exceptions.RemocaoInvalidaException;
+import sistemaDeVendas.exceptions.UpdateInvalidoException;
 import sistemaDeVendas.repositorys.ClienteRepository;
 
 @Service
@@ -13,33 +19,57 @@ public class ClienteService {
 
 	@Autowired
 	private ClienteRepository clienteRepository;
-	
-	public Cliente createCliente(Cliente c) {
-		
-		return clienteRepository.save(c);
-		
-	}
-	
-	public Cliente updateCliente(Cliente c) {
-		Cliente clienteBase = readCliente(c.getId());
-	
-		clienteBase.setEndereco(c.getEndereco());
-		clienteBase.setNome(c.getNome());
-		clienteBase.setTelefone(c.getTelefone());
 
-		return clienteRepository.save(clienteBase);
+	public Cliente createCliente(Cliente c) {
+
+		if (c == null || c.getId() != null) {
+			throw new CriacaoInvalidaException("Cliente");
+		}
+
+		return clienteRepository.save(c);
+
+	}
+
+	public Cliente updateCliente(Cliente c) {
+
+		if (c == null || c.getId() == null) {
+			throw new UpdateInvalidoException("Cliente");
+		}
+
+		return clienteRepository.save(c);
+
 	}
 
 	public Cliente readCliente(String idCliente) {
-		return clienteRepository.findById(idCliente).get();
+		
+		if (idCliente == null) {
+			throw new IdInvalidoException("id Cliente");
+		}
+
+		Optional<Cliente> c = clienteRepository.findById(idCliente);
+		
+		return c.orElseThrow(() -> NotFoundException.build("Cliente"));
+
+//		if (!c.isPresent()) {
+//			throw new NotFoundException("Cliente");
+//		}
+
+		//return c.get();
+
 	}
-	
+
 	public ArrayList<Cliente> readAll() {
 		return (ArrayList<Cliente>) clienteRepository.findAll();
 	}
-	
+
 	public Cliente removeCliente(Cliente c) {
-		clienteRepository.deleteById(c.getId());;
+
+		if (c == null || c.getId() == null) {
+			throw new RemocaoInvalidaException("Cliente");
+		}
+
+		clienteRepository.deleteById(c.getId());
+
 		return c;
 	}
 }
