+++ src/main/java/sistemaDeVendas/services/UsuarioService.java
@@ -1,11 +1,17 @@
 package sistemaDeVendas.services;
 
 import java.util.ArrayList;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import sistemaDeVendas.domains.Usuario;
+import sistemaDeVendas.exceptions.CriacaoInvalidaException;
+import sistemaDeVendas.exceptions.IdInvalidoException;
+import sistemaDeVendas.exceptions.NotFoundException;
+import sistemaDeVendas.exceptions.RemocaoInvalidaException;
+import sistemaDeVendas.exceptions.UpdateInvalidoException;
 import sistemaDeVendas.repositorys.UsuarioRepository;
 
 @Service
@@ -15,32 +21,54 @@ public class UsuarioService {
 	private UsuarioRepository usuarioRepository;
 
 	public Usuario createUsuario(Usuario u) {
-		
+
+		if (u == null || u.getId() != null) {
+			throw new CriacaoInvalidaException("Usuário");
+		}
+
 		return usuarioRepository.save(u);
-		
+
 	}
 
 	public Usuario updateUsuario(Usuario u) {
-		Usuario usuarioBase = readUsuario(u.getId());
 
-		usuarioBase.setMatriculaFuncionario(u.getMatriculaFuncionario());
-		usuarioBase.setSenha(u.getSenha());
-		
-		return usuarioRepository.save(usuarioBase);
+		if (u == null || u.getId() == null) {
+			throw new UpdateInvalidoException("Usuário");
+		}
+
+		return usuarioRepository.save(u);
 
 	}
 
 	public Usuario readUsuario(String idUsuario) {
-		return usuarioRepository.findById(idUsuario).get();
+
+		if (idUsuario == null) {
+			throw new IdInvalidoException("Id Usuário");
+		}
+
+		Optional<Usuario> u = usuarioRepository.findById(idUsuario);
+
+		return u.orElseThrow(() -> NotFoundException.build("Usuário"));
+
+//		if (!u.isPresent()) {
+//			throw new NotFoundException("Usuário");
+//		}
+
+//		return u.get();
 	}
-	
+
 	public ArrayList<Usuario> readAll() {
 		return (ArrayList<Usuario>) usuarioRepository.findAll();
 	}
-	
+
 	public Usuario removeUsuario(Usuario u) {
+
+		if (u == null || u.getId() == null) {
+			throw new RemocaoInvalidaException("Usuário");
+		}
+
 		usuarioRepository.deleteById(u.getId());
 		return u;
 	}
-	
+
 }
